# Make use of goccy/go-json for more performance increase