--- a/go.mod
+++ b/go.mod
@@ -1,16 +1,14 @@
 module github.com/ipqwery/ipapi-go
 
-go 1.23.3
+go 1.24.2
 
 require (
-	github.com/json-iterator/go v1.1.12
-	github.com/valyala/fasthttp v1.58.0
+	github.com/goccy/go-json v0.10.5
+	github.com/valyala/fasthttp v1.62.0
 )
 
 require (
 	github.com/andybalholm/brotli v1.1.1 // indirect
-	github.com/klauspost/compress v1.17.11 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/modern-go/reflect2 v1.0.2 // indirect
+	github.com/klauspost/compress v1.18.0 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
 )