Replaced `json-iterator/go` with `goccy/go-json` to gain more performance and minimize memory allocation. Additional:
- Updated go version and package to latest.
- Fixed shadowed `err` variable

### Benchmark 
Make
#### Before
```
goos: linux
goarch: amd64
pkg: github.com/ipqwery/ipapi-go
cpu: AMD Ryzen 7 5700X 8-Core Processor             
BenchmarkQueryIP-16           57          20769913 ns/op            1557 B/op         23 allocs/op
PASS
ok      github.com/ipqwery/ipapi-go     2.095s
```

#### After
```
goos: linux
goarch: amd64
pkg: github.com/ipqwery/ipapi-go
cpu: AMD Ryzen 7 5700X 8-Core Processor             
BenchmarkQueryIP-16           60          18717992 ns/op            1599 B/op          8 allocs/op
PASS
ok      github.com/ipqwery/ipapi-go     2.030s
```